feat(workflow): support limit count action
